Fix https://issues.apache.org/bugzilla/show_bug.cgi?id=57581 Change internal byte counter in Coyote Request to be long instead of int. The public API via RequestInfo.getRequestBytesReceived() already declares it as long.
Merged r1659522 from tomcat/tc8.0.x/trunk.


git-svn-id: https://svn.apache.org/repos/asf/tomcat/tc7.0.x/trunk@1659523 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/coyote/Request.java
@@ -138,7 +138,7 @@ public final class Request {
     private Response response;
     private ActionHook hook;
 
-    private int bytesRead=0;
+    private long bytesRead=0;
     // Time of the request - useful to avoid repeated calls to System.currentTime
     private long startTime = -1;
     private int available = 0;
@@ -520,7 +520,7 @@ public final class Request {
         return reqProcessorMX;
     }
 
-    public int getBytesRead() {
+    public long getBytesRead() {
         return bytesRead;
     }
